inference/requirements.txt
torch
torchvision
pandas
Pillow
opencv-python-headless
ultralytics
fastapi
uvicorn[standard]
python-multipart